_MyViewTest > moved testAddDescriptionOnMyViews, testEditDescriptionOnMyViews, testCheckButtonPreviewDescriptionOnMyViews, testCheckButtonHidePreviewDescriptionOnMyViews
Delite class FabrianaDescriptionOnMyViewsTests
